New achievements Sprite Project first clean up Various bug fixes Various graph integration

+++ Assets/Scripts/Core/Mine/MineController.cs
@@ -83,7 +83,7 @@
         {
 			ChangeResource();
             _workerNb = _db.MineStats.WorkerAmount + _db.MineUpgrades.WorkerAmount.value * _playerProfile.CurrentFortress.MineUpgradesIndex._workerNbIndex;
-			_resByWorker = _db.MineStats.ResByWorker + _db.MineUpgrades.ResByWorker.value * _playerProfile.CurrentFortress.MineUpgradesIndex._resByWorkerIndex;
+			_resByWorker = (int)(_db.MineStats.ResByWorker + _db.MineUpgrades.ResByWorker.value * _playerProfile.CurrentFortress.MineUpgradesIndex._resByWorkerIndex);
 			_luck = _db.MineStats.Luck - _db.MineUpgrades.Luck.value * _playerProfile.CurrentFortress.MineUpgradesIndex._luckIndex;
 			_richVein = _db.MineStats.RichVein + _db.MineUpgrades.RichVein.value * _playerProfile.CurrentFortress.MineUpgradesIndex._richVeinIndex;
 			_mithrilChance = _db.MineStats.Mithril - _db.MineUpgrades.Mithril.value * _playerProfile.CurrentFortress.MineUpgradesIndex._mithrilChanceIndex;
